+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * `Repos`:
     * `Common`:
         * By default, `getAll` for repos will take all the size of repo as page size
+        * New extension for all built-in repos `maxPagePagination`
+    * All the repos got `getAll` functions
 
 ## 0.17.4